Change integer field to double in Assets dto

+++ src/main/java/com/takuro_tamura/autofx/infrastructure/external/response/Assets.java
@@ -8,9 +8,9 @@ public class Assets {
     private Double availableAmount;
     private Double balance;
     private Double estimatedTradeFee;
-    private Integer margin;
+    private Double margin;
     private Double marginRatio;
     private Double positionLossGain;
     private Double totalSwap;
-    private Integer transferableAmount;
+    private Double transferableAmount;
 }
